+++ CHANGELOG.md
@@ -7,20 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.20.0]
+
+### Added
+
+- os
+  - uname, true, false commands
+  - MacOS support
+- sudo
+  - MacOS support
+- vcs
+  - MacOS support
+- arc
+  - MacOS support
+- rxtx
+  - MacOS support
+- iam
+  - MacOS support
+- pkg
+  - MacOS support
+  - Aliases and commands setter
+
+### Fixed
+
+- os
+  - Prevent shell exit when OS is not supported
+
 ## [1.19.0]
 
 ### Added
 
 - dbg
-  - add new module
+  - new module
 - core
-  - add verbose flag (BL64_LIB_VERBOSE)
+  - verbose flag (BL64_LIB_VERBOSE)
   - move debug flags to module dbg
 - xsv
   - bl64_xsv_dump: new function
   - use dbg flags
 - msg
-  - add verbose flag
+  - verbose flag
 
 ### Changed
 
@@ -156,7 +182,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - iam
   - alias for useradd
 
-[Unreleased]: https://github.com/serdigital64/bashlib64/compare/1.19.0...HEAD
+[Unreleased]: https://github.com/serdigital64/bashlib64/compare/1.20.0...HEAD
+[1.20.0]: https://github.com/serdigital64/bashlib64/compare/1.19.0...1.20.0
 [1.19.0]: https://github.com/serdigital64/bashlib64/compare/1.18.0...1.19.0
 [1.18.0]: https://github.com/serdigital64/bashlib64/compare/1.17.0...1.18.0
 [1.17.0]: https://github.com/serdigital64/bashlib64/compare/1.16.0...1.17.0